third_party/blink/web_tests/platform/mac-mac10.13/external/wpt/html/semantics/embedded-content/media-elements/track/track-element/track-cues-sorted-before-dispatch-expected.txt

This is a testharness.js-based test.
FAIL All events are triggered in chronological order assert_equals: expected 14 but got 10
Harness: the test ran to completion.
